requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
lyricsgenius==2.0.2
python-dotenv==0.20.0
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
